# Performance improvement for pagination