--- a/dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/ConverterService.java
+++ b/dspace-server-webapp/src/main/java/org/dspace/app/rest/converter/ConverterService.java
@@ -202,17 +202,18 @@ public class ConverterService {
      * @throws ClassCastException if the converter's return type is not compatible with the inferred return type.
      */
     public <M, R> Page<R> toRestPage(List<M> modelObjects, Pageable pageable, Projection projection) {
+        if (pageable == null) {
+            pageable = utils.getPageable(pageable);
+        }
+        List<M> pageableObjects = utils.getPageObjectList(modelObjects, pageable);
         List<R> transformedList = new LinkedList<>();
-        for (M modelObject : modelObjects) {
+        for (M modelObject : pageableObjects) {
             R transformedObject = toRest(modelObject, projection);
             if (transformedObject != null) {
                 transformedList.add(transformedObject);
             }
         }
-        if (pageable == null) {
-            pageable = utils.getPageable(pageable);
-        }
-        return utils.getPage(transformedList, pageable);
+        return new PageImpl(transformedList, pageable, modelObjects.size());
     }
 
     /**