@tdonohue 2023-03-23
@mspalti : Nice find, this does sound like a potentially major performance improvement.  Added to 7.6 board & we'll find some volunteer reviewers.