@mspalti 2023-03-25
This also speeds up the Item bitstreams edit panel when the item has lots of files.  So far that's the only other place where I've seen a benefit but there might be others.