@tdonohue 2023-03-30
@DSquareTechnologies: In today's dev mtg you briefly mentioned seeing performance issues for Items with many Bitstreams.  I wanted to alert you to this PR (I'm currently testing it today) as it may improve performance in that area.