@DSquareTechnologies 2023-03-30
Thanks Tim! I have forwarded this to team. Will get back in a day or two 
if they can test on volume I shared.

Narendra

On 30-03-2023 21:24, Tim Donohue wrote:
>
> @DSquareTechnologies <https://github.com/DSquareTechnologies>: In 
> today's dev mtg you briefly mentioned seeing performance issues for 
> Items with many Bitstreams. I wanted to alert you to this PR (I'm 
> currently testing it today) as it may improve performance in that area.
>
> —
> Reply to this email directly, view it on GitHub 
> <https://github.com/DSpace/DSpace/pull/8733#issuecomment-1490542904>, 
> or unsubscribe 
> <https://github.com/notifications/unsubscribe-auth/ABQHNKXBWBTWJ7YWSJWNQADW6WUBXANCNFSM6AAAAAAWFWTJXE>.
> You are receiving this because you were mentioned.Message ID: 
> ***@***.***>
>