@tdonohue 2023-03-30
@DSquareTechnologies : as of a few minutes ago, this PR was merged to `main` as it's working well for me.  So, your team can always just test the latest `main` (of both the frontend & backend) to see if the performance is improved there.  If it still isn't great, feel free to log a ticket with the exact behavior you are seeing (and ideally, if possible, steps which others can use to reproduce it)