@DSquareTechnologies 2023-03-30
Noted. Thanks
________________________________
From: Tim Donohue ***@***.***>
Sent: Thursday, March 30, 2023 10:27:30 PM
To: DSpace/DSpace ***@***.***>
Cc: DSquare Technologies ***@***.***>; Mention ***@***.***>
Subject: Re: [DSpace/DSpace] Performance improvement for pagination (PR #8733)


@DSquareTechnologies<https://github.com/DSquareTechnologies> : as of a few minutes ago, this PR was merged to main as it's working well for me. So, your team can always just test the latest main (of both the frontend & backend) to see if the performance is improved there. If it still isn't great, feel free to log a ticket with the exact behavior you are seeing (and ideally, if possible, steps which others can use to reproduce it)

—
Reply to this email directly, view it on GitHub<https://github.com/DSpace/DSpace/pull/8733#issuecomment-1490628782>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ABQHNKRBP3UZ4AQKMYP63FDW6W3PVANCNFSM6AAAAAAWFWTJXE>.
You are receiving this because you were mentioned.Message ID: ***@***.***>